requirements.txt
diffusers==0.32.2
deepspeed==0.16.5
matplotlib==3.9.0
numpy==2.2.2
peft==0.14.0
Pillow==11.0.0
PyQt5 == 5.15.10
PyYAML==6.0.2
scikit_learn==1.6.1
scipy==1.15.1
torch==2.6.0
torchrl==0.7.2
torch-fidelity
torchmetrics==1.6.1
torchvision==0.21.0
transformers==4.48.2
tqdm==4.67.1
vision-aided-loss==0.1.0